// src/common/enums/machine-type.enum.ts
export enum MachineType {
  CNC_MILL = 'CNC_MILL',
  FURNACE = 'FURNACE',
  ROBOT_ARM = 'ROBOT_ARM',
  LASER_CUTTER = 'LASER_CUTTER',
  HYDRAULIC_PRESS = 'HYDRAULIC_PRESS',
  COMPRESSOR = 'COMPRESSOR',
  CONVEYOR = 'CONVEYOR',
  PUMP = 'PUMP',
  TURBINE = 'TURBINE',
  WELDER = 'WELDER',
  THREE_D_PRINTER = 'THREE_D_PRINTER',
  INJECTION_MOLD = 'INJECTION_MOLD',
}